# api/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'College Attendance API'
  description: 'Versioned REST API for marking attendance, managing students/subjects, and generating detention & dashboard reports.'
  version: 1.0.0
servers:
  -
    url: /api/v1
    description: 'Default API server'
paths:
  /attendance:
    post:
      tags:
        - Attendance
      summary: 'Mark attendance for a class (batch)'
      operationId: 1b39ce984dd414ae463d5a8645386aae
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - subject_id
                - date
                - records
              properties:
                subject_id:
                  type: integer
                  example: 1
                date:
                  type: string
                  format: date
                  example: '2026-06-08'
                records:
                  description: 'Map of student_id => status (Present|Absent|Leave)'
                  type: object
                  example:
                    101: Present
                    102: Absent
              type: object
      responses:
        200:
          description: 'Attendance marked'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      message:
                        type: string
                      count:
                        type: integer
                    type: object
                type: object
        403:
          description: 'Not authorized for this subject'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        422:
          description: 'Validation failed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        -
          bearerAuth: []
  /attendance/students:
    get:
      tags:
        - Attendance
      summary: 'List students for marking attendance'
      operationId: dab6ab4df1f0b07caed7891712958d6d
      parameters:
        -
          name: semester
          in: query
          required: true
          schema:
            type: integer
            maximum: 8
            minimum: 1
        -
          name: branch
          in: query
          required: true
          schema:
            type: string
          example: CSE
      responses:
        200:
          description: 'List of students'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      properties:
                        student_id:
                          type: integer
                        student_name:
                          type: string
                        roll_number:
                          type: string
                      type: object
                type: object
        422:
          description: 'Validation failed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        -
          bearerAuth: []
  /attendance/subjects:
    get:
      tags:
        - Attendance
      summary: "List the authenticated faculty's subjects"
      operationId: 2ce8590b4baed077dac34fb6e7c3a3d8
      responses:
        200:
          description: 'List of subjects'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      properties:
                        subject_id:
                          type: integer
                        subject_name:
                          type: string
                        subject_code:
                          type: string
                        department:
                          type: string
                        semester:
                          type: integer
                      type: object
                type: object
      security:
        -
          bearerAuth: []
  '/attendance/monthly/{student}':
    get:
      tags:
        - Attendance
      summary: 'Get a monthly attendance summary for a student, broken down by subject'
      operationId: 13245fd5cb3d4f620e2ca7e7e5a8d0bd
      parameters:
        -
          name: student
          in: path
          required: true
          schema:
            type: integer
        -
          name: month
          in: query
          description: 'Format Y-m, defaults to last month'
          required: false
          schema:
            type: string
            format: date
            example: 2026-05
      responses:
        200:
          description: 'Monthly attendance summary'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      month:
                        type: string
                        example: 2026-05
                      total_classes:
                        type: integer
                      attended_classes:
                        type: integer
                      attendance_percentage:
                        type: number
                        format: float
                    type: object
                type: object
        404:
          description: 'Student not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    ApiError:
      properties:
        success:
          type: boolean
          example: false
        error:
          properties:
            code:
              type: string
              example: VALIDATION_FAILED
            message:
              type: string
              example: 'The given data was invalid.'
            details:
              type: object
              nullable: true
          type: object
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Enter the JWT access token returned by /api/v1/auth/login'
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []